feat: use new profile page everywhere
